requirements.txt
Django==2.2.10
django-nested-admin==3.4.0
drf-yasg==1.17.1